Restructured NodeJS code, bumped up all deprecated dependencies, changed code to avoid compatibility issues
