+++ extras/firebaseForNode.js
